move currentPlaylist tracks and userPlaylists out of App state and into local state of corresponding components
